requirements.txt
torch
torchvision
nltk
pandas
python-dotenv
pillow